pyproject.toml
[tool.poetry]
name = "pomodoro"
version = "0.1.0"
description = "pomodoro timer django backend"
authors = ["Andre Gomes <andre.gms94@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
